+++ src/expr.rs
@@ -103,14 +103,15 @@ impl TargetMatcher for target_lexicon::Triple {
         use TargetPredicate::{
             Abi, Arch, Endian, Env, Family, HasAtomic, Os, Panic, PointerWidth, Vendor,
         };
-        use target_lexicon::*;
+        use target_lexicon::{
+            self as tl, Architecture as arch, ArmArchitecture, Endianness as endian,
+            Environment as env, Mips32Architecture as mips32, Mips64Architecture as mips64,
+            OperatingSystem as os,
+        };
 
-        const NUTTX: target_lexicon::Vendor =
-            target_lexicon::Vendor::Custom(target_lexicon::CustomVendor::Static("nuttx"));
-        const RTEMS: target_lexicon::Vendor =
-            target_lexicon::Vendor::Custom(target_lexicon::CustomVendor::Static("rtems"));
-        const WALI: target_lexicon::Vendor =
-            target_lexicon::Vendor::Custom(target_lexicon::CustomVendor::Static("wali"));
+        const NUTTX: tl::Vendor = tl::Vendor::Custom(tl::CustomVendor::Static("nuttx"));
+        const RTEMS: tl::Vendor = tl::Vendor::Custom(tl::CustomVendor::Static("rtems"));
+        const WALI: tl::Vendor = tl::Vendor::Custom(tl::CustomVendor::Static("wali"));
 
         match tp {
             Abi(_) => {
@@ -119,44 +120,37 @@ impl TargetMatcher for target_lexicon::Triple {
             }
             Arch(arch) => {
                 if arch == &targ::Arch::x86 {
-                    matches!(self.architecture, Architecture::X86_32(_))
+                    matches!(self.architecture, arch::X86_32(_))
                 } else if arch == &targ::Arch::wasm32 {
-                    self.architecture == Architecture::Wasm32
-                        || self.architecture == Architecture::Asmjs
+                    self.architecture == arch::Wasm32 || self.architecture == arch::Asmjs
                 } else if arch == &targ::Arch::arm {
-                    matches!(self.architecture, Architecture::Arm(_))
+                    matches!(self.architecture, arch::Arm(_))
                 } else if arch == &targ::Arch::bpf {
-                    self.architecture == Architecture::Bpfeb
-                        || self.architecture == Architecture::Bpfel
+                    self.architecture == arch::Bpfeb || self.architecture == arch::Bpfel
                 } else if arch == &targ::Arch::x86_64 {
-                    self.architecture == Architecture::X86_64
-                        || self.architecture == Architecture::X86_64h
+                    self.architecture == arch::X86_64 || self.architecture == arch::X86_64h
                 } else if arch == &targ::Arch::mips32r6 {
                     matches!(
                         self.architecture,
-                        Architecture::Mips32(
-                            Mips32Architecture::Mipsisa32r6 | Mips32Architecture::Mipsisa32r6el
-                        )
+                        arch::Mips32(mips32::Mipsisa32r6 | mips32::Mipsisa32r6el)
                     )
                 } else if arch == &targ::Arch::mips64r6 {
                     matches!(
                         self.architecture,
-                        Architecture::Mips64(
-                            Mips64Architecture::Mipsisa64r6 | Mips64Architecture::Mipsisa64r6el
-                        )
+                        arch::Mips64(mips64::Mipsisa64r6 | mips64::Mipsisa64r6el)
                     )
                 } else if arch == &targ::Arch::amdgpu {
-                    self.architecture == Architecture::AmdGcn
+                    self.architecture == arch::AmdGcn
                 } else {
-                    match arch.0.parse::<Architecture>() {
+                    match arch.0.parse::<arch>() {
                         Ok(a) => match (self.architecture, a) {
-                            (Architecture::Aarch64(_), Architecture::Aarch64(_))
-                            | (Architecture::Mips32(_), Architecture::Mips32(_))
-                            | (Architecture::Mips64(_), Architecture::Mips64(_))
-                            | (Architecture::Powerpc64le, Architecture::Powerpc64)
-                            | (Architecture::Riscv32(_), Architecture::Riscv32(_))
-                            | (Architecture::Riscv64(_), Architecture::Riscv64(_))
-                            | (Architecture::Sparcv9, Architecture::Sparc64) => true,
+                            (arch::Aarch64(_), arch::Aarch64(_))
+                            | (arch::Mips32(_), arch::Mips32(_))
+                            | (arch::Mips64(_), arch::Mips64(_))
+                            | (arch::Powerpc64le, arch::Powerpc64)
+                            | (arch::Riscv32(_), arch::Riscv32(_))
+                            | (arch::Riscv64(_), arch::Riscv64(_))
+                            | (arch::Sparcv9, arch::Sparc64) => true,
                             (a, b) => a == b,
                         },
                         Err(_) => false,
@@ -166,8 +160,8 @@ impl TargetMatcher for target_lexicon::Triple {
             Endian(end) => match self.architecture.endianness() {
                 Ok(endian) => matches!(
                     (end, endian),
-                    (crate::targets::Endian::little, Endianness::Little)
-                        | (crate::targets::Endian::big, Endianness::Big)
+                    (crate::targets::Endian::little, endian::Little)
+                        | (crate::targets::Endian::big, endian::Big)
                 ),
 
                 Err(_) => false,
@@ -175,85 +169,82 @@ impl TargetMatcher for target_lexicon::Triple {
             Env(env) => {
                 // The environment is implied by some operating systems
                 match self.operating_system {
-                    OperatingSystem::Redox => env == &targ::Env::relibc,
-                    OperatingSystem::VxWorks => env == &targ::Env::gnu,
-                    OperatingSystem::Freebsd => env.0.is_empty(),
-                    OperatingSystem::Netbsd => match self.architecture {
-                        Architecture::Arm(ArmArchitecture::Armv6 | ArmArchitecture::Armv7) => {
+                    os::Redox => env == &targ::Env::relibc,
+                    os::VxWorks => env == &targ::Env::gnu,
+                    os::Freebsd => env.0.is_empty(),
+                    os::Netbsd => match self.architecture {
+                        arch::Arm(ArmArchitecture::Armv6 | ArmArchitecture::Armv7) => {
                             env.0.is_empty()
                         }
                         _ => env.0.is_empty(),
                     },
-                    OperatingSystem::None_
-                    | OperatingSystem::Cloudabi
-                    | OperatingSystem::Hermit
-                    | OperatingSystem::IOS(_) => match self.environment {
-                        Environment::LinuxKernel => env == &targ::Env::gnu,
+                    os::None_ | os::Cloudabi | os::Hermit => match self.environment {
+                        env::LinuxKernel => env == &targ::Env::gnu,
                         _ => env.0.is_empty(),
                     },
-                    OperatingSystem::WasiP1 => env == &targ::Env::p1,
-                    OperatingSystem::WasiP2 => env == &targ::Env::p2,
-                    OperatingSystem::Wasi => env.0.is_empty() || env == &targ::Env::p1,
+                    os::IOS(_) | os::TvOS(_) => match self.environment {
+                        env::LinuxKernel => env == &targ::Env::gnu,
+                        env::Macabi => env == &targ::Env::macabi,
+                        env::Sim => env == &targ::Env::sim,
+                        env::Unknown => env.0.is_empty() || env == &targ::Env::sim,
+                        _ => env.0.is_empty(),
+                    },
+                    os::WasiP1 => env == &targ::Env::p1,
+                    os::WasiP2 => env == &targ::Env::p2,
+                    os::Wasi => env.0.is_empty() || env == &targ::Env::p1,
                     _ => {
                         if env.0.is_empty() {
                             matches!(
                                 self.environment,
-                                Environment::Unknown
-                                    | Environment::Android
-                                    | Environment::Softfloat
-                                    | Environment::Androideabi
-                                    | Environment::Eabi
-                                    | Environment::Eabihf
-                                    | Environment::Sim
-                                    | Environment::None
+                                env::Unknown
+                                    | env::Android
+                                    | env::Softfloat
+                                    | env::Androideabi
+                                    | env::Eabi
+                                    | env::Eabihf
+                                    | env::Sim
+                                    | env::None
                             )
                         } else {
-                            match env.0.parse::<Environment>() {
+                            match env.0.parse::<env>() {
                                 Ok(e) => {
                                     // Rustc shortens multiple "gnu*" environments to just "gnu"
                                     if env == &targ::Env::gnu {
                                         match self.environment {
-                                            Environment::Gnu
-                                            | Environment::Gnuabi64
-                                            | Environment::Gnueabi
-                                            | Environment::Gnuspe
-                                            | Environment::Gnux32
-                                            | Environment::GnuIlp32
-                                            | Environment::Gnueabihf
-                                            | Environment::GnuLlvm => true,
+                                            env::Gnu
+                                            | env::Gnuabi64
+                                            | env::Gnueabi
+                                            | env::Gnuspe
+                                            | env::Gnux32
+                                            | env::GnuIlp32
+                                            | env::Gnueabihf
+                                            | env::GnuLlvm => true,
                                             // Rust 1.49.0 changed all android targets to have the
                                             // gnu environment
-                                            Environment::Android | Environment::Androideabi
-                                                if self.operating_system
-                                                    == OperatingSystem::Linux =>
+                                            env::Android | env::Androideabi
+                                                if self.operating_system == os::Linux =>
                                             {
                                                 true
                                             }
-                                            Environment::Kernel => {
-                                                self.operating_system == OperatingSystem::Linux
-                                            }
-                                            _ => self.architecture == Architecture::Avr,
+                                            env::Kernel => self.operating_system == os::Linux,
+                                            _ => self.architecture == arch::Avr,
                                         }
                                     } else if env == &targ::Env::musl {
                                         matches!(
                                             self.environment,
-                                            Environment::Musl
-                                                | Environment::Musleabi
-                                                | Environment::Musleabihf
-                                                | Environment::Muslabi64
+                                            env::Musl
+                                                | env::Musleabi
+                                                | env::Musleabihf
+                                                | env::Muslabi64
                                         )
                                     } else if env == &targ::Env::uclibc {
                                         matches!(
                                             self.environment,
-                                            Environment::Uclibc
-                                                | Environment::Uclibceabi
-                                                | Environment::Uclibceabihf
+                                            env::Uclibc | env::Uclibceabi | env::Uclibceabihf
                                         )
                                     } else if env == &targ::Env::newlib {
-                                        matches!(
-                                            self.operating_system,
-                                            OperatingSystem::Horizon | OperatingSystem::Espidf
-                                        ) || self.vendor == RTEMS
+                                        matches!(self.operating_system, os::Horizon | os::Espidf)
+                                            || self.vendor == RTEMS
                                     } else {
                                         self.environment == e
                                     }
@@ -265,71 +256,72 @@ impl TargetMatcher for target_lexicon::Triple {
                 }
             }
             Family(fam) => {
-                use OperatingSystem::{
-                    Aix, AmdHsa, Bitrig, Cloudabi, Cuda, Darwin, Dragonfly, Emscripten, Espidf,
-                    Freebsd, Fuchsia, Haiku, Hermit, Horizon, Hurd, IOS, Illumos, L4re, Linux,
-                    MacOSX, Nebulet, Netbsd, None_, Openbsd, Redox, Solaris, TvOS, Uefi, Unknown,
-                    VisionOS, VxWorks, Wasi, WasiP1, WasiP2, WatchOS, Windows,
-                };
-
                 match self.operating_system {
-                    AmdHsa | Bitrig | Cloudabi | Cuda | Hermit | Nebulet | None_ | Uefi => false,
-                    Aix
-                    | Darwin(_)
-                    | Dragonfly
-                    | Espidf
-                    | Freebsd
-                    | Fuchsia
-                    | Haiku
-                    | Hurd
-                    | Illumos
-                    | IOS(_)
-                    | L4re
-                    | MacOSX { .. }
-                    | Horizon
-                    | Netbsd
-                    | Openbsd
-                    | Redox
-                    | Solaris
-                    | TvOS(_)
-                    | VisionOS(_)
-                    | VxWorks
-                    | WatchOS(_) => fam == &crate::targets::Family::unix,
-                    Emscripten => {
+                    os::AmdHsa
+                    | os::Bitrig
+                    | os::Cloudabi
+                    | os::Cuda
+                    | os::Hermit
+                    | os::Nebulet
+                    | os::None_
+                    | os::Uefi => false,
+                    os::Aix
+                    | os::Darwin(_)
+                    | os::Dragonfly
+                    | os::Espidf
+                    | os::Freebsd
+                    | os::Fuchsia
+                    | os::Haiku
+                    | os::Hurd
+                    | os::Illumos
+                    | os::IOS(_)
+                    | os::L4re
+                    | os::MacOSX { .. }
+                    | os::Horizon
+                    | os::Netbsd
+                    | os::Openbsd
+                    | os::Redox
+                    | os::Solaris
+                    | os::TvOS(_)
+                    | os::VisionOS(_)
+                    | os::VxWorks
+                    | os::WatchOS(_) => fam == &crate::targets::Family::unix,
+                    os::Emscripten => {
                         match self.architecture {
                             // asmjs, wasm32 and wasm64 are part of both the wasm and unix families
-                            Architecture::Asmjs | Architecture::Wasm32 => {
+                            arch::Asmjs | arch::Wasm32 => {
                                 fam == &crate::targets::Family::wasm
                                     || fam == &crate::targets::Family::unix
                             }
                             _ => false,
                         }
                     }
-                    Unknown if self.vendor == NUTTX || self.vendor == RTEMS => {
+                    os::Unknown if self.vendor == NUTTX || self.vendor == RTEMS => {
                         fam == &crate::targets::Family::unix
                     }
-                    Unknown => {
+                    os::Unknown => {
                         // asmjs, wasm32 and wasm64 are part of the wasm family.
                         match self.architecture {
-                            Architecture::Asmjs | Architecture::Wasm32 | Architecture::Wasm64 => {
+                            arch::Asmjs | arch::Wasm32 | arch::Wasm64 => {
                                 fam == &crate::targets::Family::wasm
                             }
                             _ => false,
                         }
                     }
-                    Linux if self.vendor == WALI => {
+                    os::Linux if self.vendor == WALI => {
                         fam == &crate::targets::Family::wasm || fam == &crate::targets::Family::unix
                     }
-                    Linux => {
+                    os::Linux => {
                         // The 'kernel' environment is treated specially as not-unix
-                        if self.environment != Environment::Kernel {
+                        if self.environment != env::Kernel {
                             fam == &crate::targets::Family::unix
                         } else {
                             false
                         }
                     }
-                    Wasi | WasiP1 | WasiP2 => fam == &crate::targets::Family::wasm,
-                    Windows => fam == &crate::targets::Family::windows,
+                    os::Wasi | os::WasiP1 | os::WasiP2 => fam == &crate::targets::Family::wasm,
+                    os::Windows => fam == &crate::targets::Family::windows,
+                    os::Cygwin => fam == &crate::targets::Family::unix,
                     // I really dislike non-exhaustive :(
                     _ => false,
                 }
@@ -340,35 +332,30 @@ impl TargetMatcher for target_lexicon::Triple {
                 false
             }
             Os(os) => {
-                if os == &targ::Os::wasi
-                    && matches!(
-                        self.operating_system,
-                        OperatingSystem::WasiP1 | OperatingSystem::WasiP2
-                    )
+                if os == &targ::Os::wasi && matches!(self.operating_system, os::WasiP1 | os::WasiP2)
                     || (os == &targ::Os::nuttx && self.vendor == NUTTX)
                     || (os == &targ::Os::rtems && self.vendor == RTEMS)
                 {
                     return true;
                 }
 
-                match os.0.parse::<OperatingSystem>() {
+                match os.0.parse::<os>() {
                     Ok(o) => match self.environment {
-                        Environment::HermitKernel => os == &targ::Os::hermit,
+                        env::HermitKernel => os == &targ::Os::hermit,
                         _ => self.operating_system == o,
                     },
                     Err(_) => {
                         // Handle special case for darwin/macos, where the triple is
                         // "darwin", but rustc identifies the OS as "macos"
-                        if os == &targ::Os::macos
-                            && matches!(self.operating_system, OperatingSystem::Darwin(_))
+                        if os == &targ::Os::macos && matches!(self.operating_system, os::Darwin(_))
                         {
                             true
                         } else {
                             // For android, the os is still linux, but the environment is android
                             os == &targ::Os::android
-                                && self.operating_system == OperatingSystem::Linux
-                                && (self.environment == Environment::Android
-                                    || self.environment == Environment::Androideabi)
+                                && self.operating_system == os::Linux
+                                && (self.environment == env::Android
+                                    || self.environment == env::Androideabi)
                         }
                     }
                 }
@@ -384,10 +371,9 @@ impl TargetMatcher for target_lexicon::Triple {
                             && ven == &targ::Vendor::unknown)
                     {
                         true
-                    } else if let target_lexicon::Vendor::Custom(custom) = &self.vendor {
+                    } else if let tl::Vendor::Custom(custom) = &self.vendor {
                         matches!(custom.as_str(), "esp" | "esp32" | "esp32s2" | "esp32s3")
-                            && (v == target_lexicon::Vendor::Espressif
-                                || v == target_lexicon::Vendor::Unknown)
+                            && (v == tl::Vendor::Espressif || v == tl::Vendor::Unknown)
                     } else {
                         false
                     }
@@ -397,10 +383,7 @@ impl TargetMatcher for target_lexicon::Triple {
             PointerWidth(pw) => {
                 // The gnux32 environment is a special case, where it has an
                 // x86_64 architecture, but a 32-bit pointer width
-                if !matches!(
-                    self.environment,
-                    Environment::Gnux32 | Environment::GnuIlp32
-                ) {
+                if !matches!(self.environment, env::Gnux32 | env::GnuIlp32) {
                     *pw == match self.pointer_width() {
                         Ok(pw) => pw.bits(),
                         Err(_) => return false,